Reorganize settings
