+++ backend/backend/settings/deps/__init__.py
@@ -1,0 +1,3 @@
+"""
+self-contained categorized settings
+"""
