+++ backend/backend/settings/deps/jwt.py
@@ -1,22 +1,17 @@
+"""
+JSON web tokens
+"""
+
 import os
-import pathlib
-import secrets
+from datetime import timedelta
 
-from .. import common
+SIMPLE_JWT = {
+    "ACCESS_TOKEN_LIFETIME": timedelta(minutes=int(os.environ.get("ACCESS_TOKEN_LIFETIME", 24 * 60))),
+    "REFRESH_TOKEN_LIFETIME": timedelta(minutes=int(os.environ.get("REFRESH_TOKEN_LIFETIME", 24 * 60 * 7))),
+    "ROTATE_REFRESH_TOKENS": True,
+    "AUTH_HEADER_TYPES": ("JWT",),
+    "BLACKLIST_AFTER_ROTATION": True,
+}
 
-# SECURITY WARNING: keep the secret key used in production secret!
-JWT_SECRET_PATH = os.environ.get("JWT_SECRET_PATH", os.path.normpath(os.path.join(common.PROJECT_ROOT, "SECRET")))
-
-# Key configuration for JSON web tokens (JWT) authentication
-if common.to_bool(os.environ.get("JWT_SECRET_NEEDED", "False")):
-    try:
-        JWT_SECRET_KEY = pathlib.Path(JWT_SECRET_PATH).read_text().strip()
-    except IOError:
-        try:
-            JWT_SECRET_KEY = secrets.token_urlsafe()  # uses a "reasonable default" length
-            with open(JWT_SECRET_PATH, "w") as fp:
-                fp.write(JWT_SECRET_KEY)
-        except IOError:
-            raise Exception(f"Cannot open file `{JWT_SECRET_PATH}` for writing.")
-else:
-    JWT_SECRET_KEY = "unused default value " + secrets.token_urlsafe()
+# To encode unique jwt token generated with reset password request
+RESET_JWT_SIGNATURE_ALGORITHM = "HS256"
