+++ backend/backend/settings/deps/path.py
@@ -1,0 +1,9 @@
+import os
+import sys
+
+# Build paths inside the project like this: os.path.join(BASE_DIR, ...)
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+PROJECT_ROOT = os.path.dirname(BASE_DIR)
+
+sys.path.append(PROJECT_ROOT)
+sys.path.append(os.path.normpath(os.path.join(PROJECT_ROOT, "libs")))
