+++ backend/backend/settings/dev.py
@@ -3,11 +3,11 @@ import os
 from substrapp.storages.minio import MinioStorage
 
 from .common import *
-from .deps.cors import *
 from .deps.ledger import *
-from .deps.oidc import *
 from .deps.org import *
 from .deps.restframework import *
+from .mods.cors import *
+from .mods.oidc import *
 
 DEBUG = True
