+++ backend/backend/settings/mods/cors.py
@@ -1,7 +1,12 @@
+"""
+Add Cross-origin site requests (CORS) restrictions
+"""
+
 import json
 import os
 
 from .. import common
+from ..deps.utils import to_bool
 
 common.INSTALLED_APPS += ("corsheaders",)
 
@@ -11,7 +16,7 @@ common.MIDDLEWARE.insert(0, "corsheaders.middleware.CorsMiddleware")
 CORS_ALLOWED_ORIGINS = json.loads(os.environ.get("CORS_ORIGIN_WHITELIST", "[]"))
 CORS_ALLOW_ALL_ORIGINS = False
 # @CORS_ALLOW_CREDENTIALS: If True cookies can be included in cross site requests. Set this to `True` for frontend auth.
-CORS_ALLOW_CREDENTIALS = common.to_bool(os.environ.get("CORS_ALLOW_CREDENTIALS", False))
+CORS_ALLOW_CREDENTIALS = to_bool(os.environ.get("CORS_ALLOW_CREDENTIALS", False))
 CORS_ALLOW_HEADERS = (
     "accept",
     "accept-encoding",
