+++ backend/backend/settings/mods/oidc.py
@@ -1,10 +1,15 @@
+"""
+OpenID Connect (for SSO)
+"""
+
 import logging
 import os
 
 import requests
 
 from .. import common
-from . import ledger
+from ..deps import ledger
+from ..deps.utils import to_bool
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -13,7 +18,7 @@ _LOGGER = logging.getLogger(__name__)
 # we'll use an "OIDC" dict for any setting added by us
 
 OIDC = {
-    "ENABLED": common.to_bool(os.environ.get("OIDC_ENABLED", "false")),
+    "ENABLED": to_bool(os.environ.get("OIDC_ENABLED", "false")),
     "USERS": {},
     "OP": {},
 }
@@ -26,7 +31,7 @@ if OIDC["ENABLED"]:  # noqa: C901
         "propagate": False,
     }
 
-    OIDC["USERS"]["APPEND_DOMAIN"] = common.to_bool(os.environ.get("OIDC_USERS_APPEND_DOMAIN", "false"))
+    OIDC["USERS"]["APPEND_DOMAIN"] = to_bool(os.environ.get("OIDC_USERS_APPEND_DOMAIN", "false"))
 
     OIDC["USERS"]["DEFAULT_CHANNEL"] = os.environ.get("OIDC_USERS_DEFAULT_CHANNEL")
     if not OIDC["USERS"]["DEFAULT_CHANNEL"]:
@@ -53,7 +58,7 @@ if OIDC["ENABLED"]:  # noqa: C901
     OIDC["OP"]["DISPLAY_NAME"] = os.environ.get("OIDC_OP_DISPLAY_NAME", OIDC["OP"]["URL"])
 
     OIDC["OP"]["SUPPORTS_REFRESH"] = False
-    OIDC["USERS"]["USE_REFRESH_TOKEN"] = common.to_bool(os.environ.get("OIDC_USERS_USE_REFRESH_TOKEN", "false"))
+    OIDC["USERS"]["USE_REFRESH_TOKEN"] = to_bool(os.environ.get("OIDC_USERS_USE_REFRESH_TOKEN", "false"))
 
     op_settings = None
     try:
